requirements.txt
tqdm==4.67.1
torch>=2.6.0
peft==0.17.0
openai==2.8.0
einops==0.8.0
loguru==0.7.3
numpy==1.26.4
imageio==2.37.0
angelslim==0.2.1
imageio-ffmpeg==0.6.0
omegaconf>=2.3.0
torchaudio==2.6.0
diffusers==0.32.2
safetensors>=0.5.3
torchvision>=0.21.0
qwen-vl-utils==0.0.8
huggingface-hub==0.34.0
huggingface_hub[cli]
transformers[accelerate,tiktoken]==4.57.1
pillow
wheel
psutil
setuptools
modelscope